+++ src/main/kotlin/dev/aaronhowser/mods/aaron/block_walker/ConnectedBlock.kt
@@ -1,0 +1,6 @@
+package dev.aaronhowser.mods.aaron.block_walker
+
+class ConnectedBlock(
+	val block: PositionedBlock,
+	val distance: Int
+)