+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Update the recipe provider
 - Moved some classes around
 - IntrinsicHolderTagsProvider$IntrinsicTagAppender#add(vararg values: Holder)
+- Block Walker stuff
 
 # 1.5.3